+++ convert-junit4-to-junit5/src/test/java/jb/convert/ast/AssertMigration_FailMessagePositionTest.java
@@ -1,7 +1,5 @@
-package jb;
+package jb.convert.ast;
 
-import org.junit.Assert;
-import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
@@ -9,7 +7,7 @@ import org.junit.jupiter.params.provider.ValueSource;
 import static jb.JunitConversionLogicFixture.assertAfterWrappingInMethod;
 import static jb.JunitConversionLogicFixture.assertUnchangedAfterWrappingInMethod;
 
-class MoveAssertionMessageTest {
+class AssertMigration_FailMessagePositionTest {
 
     @Test
     void notAnAssertion() {
